# Frontend/ResultCollector/OpenAPI/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "WELLSHIP",
    "description": "WELLSHIPのバックエンドです。",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://10.191.26.49/wellship"
    }
  ],
  "paths": {
...
    "/api/v{version}/consult/consultNumber/verify": {
      "post": {
        "tags": [
          "Consult"
        ],
        "summary": "受診番号の存在を確認する",
        "operationId": "Consult_VerifyConsultNumber",
        "parameters": [
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "consultNumberRequest",
          "description": "受診番号リクエスト",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultNumberRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "存在するか"
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/consult/{consultNumber}/unexaminedMenus": {
      "get": {
        "tags": [
          "Consult"
        ],
        "summary": "未受診の検査メニューを取得する",
        "operationId": "Consult_GetUnexaminedMenus",
        "parameters": [
          {
            "name": "consultNumber",
            "in": "path",
            "required": true,
            "description": "受診番号",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnexaminedMenuList"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/consult/{consultNumber}/examItems": {
      "get": {
        "tags": [
          "Consult"
        ],
        "summary": "検査内容を取得する",
        "operationId": "Consult_GetExamItemsExaminee",
        "parameters": [
          {
            "name": "consultNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "examMenuId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamContent"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/consult/{consultNumber}/executions": {
      "post": {
        "tags": [
          "Consult"
        ],
        "summary": "検査の実施有無と中止理由を登録する",
        "operationId": "Consult_RegisterExecutions",
        "parameters": [
          {
            "name": "consultNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "executions",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionsRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/consult/{consultNumber}/inputExamItems": {
      "get": {
        "tags": [
          "Consult"
        ],
        "summary": "検査結果入力情報を取得する",
        "operationId": "Consult_GetInputExamItemsExaminee",
        "parameters": [
          {
            "name": "consultNumber",
            "in": "path",
            "required": true,
            "description": "受診番号",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "examMenuId",
            "in": "query",
            "description": "検査メニューID",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputExamItems"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/consult/{consultNumber}/prior/{currentExamId}": {
      "get": {
        "tags": [
          "Consult"
        ],
        "summary": "前提検査メニューを検証する（仮：動作確認用）",
        "operationId": "Consult_ValidatePriorExamMenus",
        "parameters": [
          {
            "name": "consultNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "currentExamId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputExamItems"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/consult/{consultNumber}/correlation": {
      "post": {
        "tags": [
          "Consult"
        ],
        "summary": "検査結果相関ルールを検証する（仮：動作確認用）",
        "operationId": "Consult_ValidateCorrelationRule",
        "parameters": [
          {
            "name": "consultNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultsRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v{version}/consult/{consultNumber}/results": {
      "post": {
        "tags": [
          "Consult"
        ],
        "summary": "検査結果を登録する",
        "operationId": "Consult_RegisterResults",
        "parameters": [
          {
            "name": "consultNumber",
            "in": "path",
            "required": true,
            "description": "受診番号",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "results",
          "description": "検査結果登録項目",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultsRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/consult/{consultNumber}/results/verify": {
      "post": {
        "tags": [
          "Consult"
        ],
        "summary": "検査結果を検証する",
        "operationId": "Consult_VerifyResults",
        "parameters": [
          {
            "name": "consultNumber",
            "in": "path",
            "required": true,
            "description": "受診番号",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 3
          }
        ],
        "requestBody": {
          "x-name": "results",
          "description": "検査結果検証項目",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultsRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyExamItems"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ConsultNumberRequest": {
        "type": "object",
        "description": "受診番号検証用リクエストモデル",
        "additionalProperties": false,
        "properties": {
          "consultNumber": {
            "type": "string",
            "description": "受診番号"
          }
        }
      },
      "UnexaminedMenuList": {
        "type": "object",
        "description": "未受診の検査項目リスト",
        "additionalProperties": false,
        "properties": {
          "consultId": {
            "type": "string",
            "description": "受診ID",
            "format": "guid"
          },
          "examineeId": {
            "type": "string",
            "description": "受診者ID",
            "format": "guid"
          },
          "examineeName": {
            "type": "string",
            "description": "受診者名"
          },
          "unexaminedMenus": {
            "type": "array",
            "description": "未受診の検査メニューリスト",
            "items": {
              "$ref": "#/components/schemas/ExamMenu"
            }
          }
        }
      },
      "ExamMenu": {
        "type": "object",
        "description": "検査メニュー",
        "additionalProperties": false,
        "properties": {
          "examMenuId": {
            "type": "integer",
            "description": "検査メニューID",
            "format": "int32"
          },
          "examMenuName": {
            "type": "string",
            "description": "検査メニュー名"
          }
        }
      },
      "ExamContent": {
        "type": "object",
        "description": "検査内容",
        "additionalProperties": false,
        "properties": {
          "consultNumber": {
            "type": "string",
            "description": "受診番号"
          },
          "consultName": {
            "type": "string",
            "description": "特記事項"
          },
          "examinee": {
            "description": "受診者情報",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Examinee"
              }
            ]
          },
          "isComplete": {
            "type": "boolean",
            "description": "検査実施判断"
          },
          "relatedExamItems": {
            "type": "array",
            "description": "関連検査項目",
            "items": {
              "$ref": "#/components/schemas/RelatedExamItem"
            }
          },
          "examItems": {
            "type": "array",
            "description": "実施検査項目",
            "items": {
              "$ref": "#/components/schemas/ExamDetail"
            }
          },
          "examDecisionResult": {
            "type": "array",
            "description": "検査実施判断結果",
            "items": {
              "type": "string"
            }
          },
          "unexaminedItems": {
            "type": "array",
            "description": "未実施検査項目",
            "items": {
              "$ref": "#/components/schemas/ExamMenu"
            }
          }
        }
      },
      "Examinee": {
        "type": "object",
        "description": "受診者情報",
        "additionalProperties": false,
        "properties": {
          "ticketNumber": {
            "type": "string",
            "description": "受付番号"
          },
          "name": {
            "type": "string",
            "description": "氏名"
          },
          "kanaName": {
            "type": "string",
            "description": "カナ氏名"
          },
          "birthdate": {
            "type": "string",
            "description": "生年月日",
            "format": "date"
          },
          "sex": {
            "type": "integer",
            "description": "性別",
            "format": "int32"
          },
          "organizations": {
            "type": "array",
            "description": "事業所名",
            "items": {
              "type": "string"
            }
          },
          "sameNameAlert": {
            "type": "boolean",
            "description": "同姓同名アラート"
          },
          "examDateAge": {
            "type": "integer",
            "description": "受診日年齢",
            "format": "int32"
          }
        }
      },
      "RelatedExamItem": {
        "type": "object",
        "description": "関連検査項目",
        "additionalProperties": false,
        "properties": {
          "examItemName": {
            "type": "string",
            "description": "検査項目名"
          },
          "examResult": {
            "type": "string",
            "description": "検査結果"
          }
        }
      },
      "ExamDetail": {
        "type": "object",
        "description": "実施検査状況",
        "additionalProperties": false,
        "properties": {
          "examItemId": {
            "type": "integer",
            "description": "検査項目ID",
            "format": "int32"
          },
          "examItemName": {
            "type": "string",
            "description": "検査項目名"
          },
          "hasOrder": {
            "type": "boolean",
            "description": "検査依頼が存在するか"
          },
          "cancelReasonId": {
            "type": "integer",
            "description": "中止理由ID",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "ExecutionsRequest": {
        "type": "object",
        "description": "実施有無と中止理由を束ねたリクエストモデル",
        "additionalProperties": false,
        "properties": {
          "executions": {
            "type": "array",
            "description": "検査項目ごとの実施有無と中止理由",
            "items": {
              "$ref": "#/components/schemas/ExecutionRequest"
            }
          }
        }
      },
      "ExecutionRequest": {
        "type": "object",
        "description": "実施有無と中止理由のリクエストモデル",
        "additionalProperties": false,
        "properties": {
          "examItemId": {
            "type": "integer",
            "description": "検査項目ID",
            "format": "int32"
          },
          "isPerforming": {
            "type": "boolean",
            "description": "検査実施するか"
          },
          "cancelReasonId": {
            "type": "integer",
            "description": "中止理由ID",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "InputExamItems": {
        "type": "object",
        "description": "検査結果入力情報",
        "additionalProperties": false,
        "properties": {
          "consultNumber": {
            "type": "string",
            "description": "受診番号"
          },
          "examinee": {
            "description": "受診者情報",
            "oneOf": [
              {
                "$ref": "#/components/schemas/InputExamExaminee"
              }
            ]
          },
          "relatedExamItems": {
            "type": "array",
            "description": "関連検査項目",
            "items": {
              "$ref": "#/components/schemas/RelatedExamItem"
            }
          },
          "examItemGroups": {
            "type": "array",
            "description": "検査結果入力項目グループ",
            "items": {
              "$ref": "#/components/schemas/ExamItemGroup"
            }
          }
        }
      },
      "InputExamExaminee": {
        "type": "object",
        "description": "検査結果入力画面 受診者情報",
        "additionalProperties": false,
        "properties": {
          "ticketNumber": {
            "type": "string",
            "description": "受付番号"
          },
          "kanaName": {
            "type": "string",
            "description": "カナ氏名"
          },
          "sex": {
            "type": "integer",
            "description": "性別",
            "format": "int32"
          },
          "examDateAge": {
            "type": "integer",
            "description": "受診日年齢",
            "format": "int32"
          }
        }
      },
      "ExamItemGroup": {
        "type": "object",
        "description": "検査項目グループ",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "integer",
            "description": "検査項目グループ種別",
            "format": "int32"
          },
          "examItems": {
            "type": "array",
            "description": "検査結果入力項目",
            "items": {
              "$ref": "#/components/schemas/InputExamItem"
            }
          }
        }
      },
      "InputExamItem": {
        "type": "object",
        "description": "検査結果入力項目",
        "additionalProperties": false,
        "properties": {
          "positionNumber": {
            "type": "integer",
            "description": "配置番号",
            "format": "int32"
          },
          "examItemId": {
            "type": "integer",
            "description": "検査項目ID",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "検査項目名"
          },
          "examItemDetails": {
            "type": "array",
            "description": "検査項目明細",
            "items": {
              "$ref": "#/components/schemas/ExamItemDetail"
            }
          },
          "examRegistResults": {
            "type": "array",
            "description": "検査結果登録エラー",
            "items": {
              "$ref": "#/components/schemas/ExamRegistResult"
            }
          }
        }
      },
...
      "ResultsRequest": {
        "type": "object",
        "description": "検査結果の登録のリクエストモデル",
        "additionalProperties": false,
        "properties": {
          "examMenuId": {
            "type": "integer",
            "description": "検査メニューID",
            "format": "int32"
          },
          "examResults": {
            "type": "array",
            "description": "検査項目",
            "items": {
              "$ref": "#/components/schemas/ResultRequest"
            }
          }
        }
      },
      "ResultRequest": {
        "type": "object",
        "description": "検査結果の登録の検査項目リクエストモデル",
        "additionalProperties": false,
        "properties": {
          "examItemId": {
            "type": "integer",
            "description": "検査項目ID",
            "format": "int32"
          },
          "examItemDetails": {
            "type": "array",
            "description": "検査項目明細",
            "items": {
              "$ref": "#/components/schemas/ExamItemDetailRequest"
            }
          }
        }
      },
      "ExamItemDetailRequest": {
        "type": "object",
        "description": "検査結果の登録の検査項目明細リクエストモデル",
        "additionalProperties": false,
        "properties": {
          "examItemDetailId": {
            "type": "integer",
            "description": "検査項目明細ID",
            "format": "int32"
          },
          "value": {
            "type": "string",
            "description": "値"
          }
        }
      },
      "VerifyExamItems": {
        "type": "object",
        "description": "検査結果検証結果情報",
        "additionalProperties": false,
        "properties": {
          "errorLevel": {
            "type": "integer",
            "description": "エラーレベル",
            "format": "int32"
          },
          "examItemGroups": {
            "type": "array",
            "description": "検査結果入力項目グループ",
            "items": {
              "$ref": "#/components/schemas/ExamItemGroup"
            }
          }
        }
      },
...
    }
  }
}